setup.py: Install the script as 'scripts', rather than 'data_files'
Install the Python script using the 'scripts' key rather than
the 'data_files' hack. This ensures that it is installed
in the correct location (respecting install options) and its shebang
is updated to match the correct Python interpreter.

This is the same patch as I've sent back in 2012, rebased on top
of the newest release.